Update README and some file renaming
