+++ high_level/common/ffg.py
@@ -3,7 +3,7 @@ from pyrsistent import PSet
 from .data_structures import *
 from .formal_verification_annotations import *
 from .pythonic_code_generic import *
-from .common import *
+from .helpers import *
 from .stubs import *
